First version working, more refactoring required
This has some logic extracted out of the Account object (which I
consider in this case to essentially be the controller), but I've left in
for the moment the logic for the statement object.

However this definitely needs to be refactored, some code smells:

- I really don't think it's the account's responsibility to know how
  to create a statement
- Given this class is just a controller/interface class, this is way too
  much logic
- The method is way too long
- Those private methods at the bottom are an indicator that we're trying
  to smuggle another object into our Account object, this will come to
bite us as the complexity of this app increases.