Extract statement logic into its own class
This is looking better, we've now cleaned up our Account object so it's
transformed into a boring, small object which does nothing more than
delegate all of its logic to other classes (which is generally what I
want my controller class to do).

However just copying the code over into another class is not enough. We
still have an enormous method, and those weird private formatting
methods still exist that don't really feel like the responsibility of a
statement object.

The fact that it we keep having to wrap those format
methods around our transaction object looks messy and is another
indicator that we might not have got the design quite right yet. At this
point it seems worth considering that it is likely that this pattern of
needing some sort of object that wraps around another object and handles
how to do all the formatting of the object it wraps around is pretty
common.

If in programming something is pretty common, it's likely there will be
a [design pattern](https://en.wikipedia.org/wiki/Software_design_pattern)
that can be used to solve the problem you're trying to solve.

In this case have a read about the Presenter pattern, I'll implement it
in the next commit.