Clean up statement view
It may have seemed that the last step wasn't worth it, and creating that
extra object was not worth the hassle - after all you now have more
classes to worry about, more tests to write, and so on.

Similarly we could have stopped at the last step and said that our
Statement object was clean enough. But if possible, going that extra
step is the difference in the long-run between a clean, maintainable
codebase and one that becomes a mess of technical debt.

Here we see the pay-off of thinking about our design properly. By
extracting the formatting into a Presenter class, we can now clean up
the Statement class massively. Not only is the code cleaner, but by
being able to extract the headers into a constant at the top, if we
decide we need to add an extra column to our bank statement changing
this code will be trivial, a sign of code that's well designed.